Tinkered with the coordinate system and figured it out.
The problem with my previous commit's experiments to understand Vulkan's NDC coordinate system was face culling direction. One of the projection matrix's Y axes could be flipped, or the camera matrix's "up" axis could be flipped to put it on the other side of the triangle faces, or face culling could be switched from counterclockwise to clockwise, or the vertices themselvs could be rearranged, and any of them could have fixed the blank screen. Pick two of the above to change, and you'll see the same result (I think).