Got it working. I forgot to change the type in vkCmdBindIndexBuffer(...) from uint_16 -> uint_32.
The tutorial said to do it, and I was going to, and I wanted to see what would happen if I didn't, and now I will remember :). Using a uint_16 to navigate a buffer of 32bit index values that range up to 1.5 million will see (1) a lot of wrap-around (2^16 maxes out around 65k) and (2) two index values read per actual 32bit index. Whoops.